Fix smt-generation bug
Rather than always cutting N assignments after exiting a context that
quantified N variables, smt-generation would always cut 1 assignment.
This was caused by iterating over [0..n], which is a singleton array
containing a range, rather than 0..n, which is a range of length N.

When an N=0 quantifier is used, this results in cutting an assignment
that should still be in scope.